Undo move of UnsignedSsvMessage for now
