+++ anchor/common/qbft/src/tests.rs
@@ -10,7 +10,7 @@ use std::{
 
 use qbft_types::DefaultLeaderFunction;
 use sha2::{Digest, Sha256};
-use ssv_types::{OperatorId, RSA_SIGNATURE_SIZE, signed_message::SignedSSVMessage};
+use ssv_types::{OperatorId, RSA_SIGNATURE_SIZE, message::SignedSSVMessage};
 use ssz_derive::{Decode, Encode};
 use tracing::debug_span;
 use tracing_subscriber::filter::EnvFilter;
