+++ anchor/message_sender/src/testing.rs
@@ -1,6 +1,6 @@
 use ssv_types::{
     CommitteeId, OperatorId, RSA_SIGNATURE_SIZE, consensus::UnsignedSSVMessage,
-    signed_message::SignedSSVMessage,
+    message::SignedSSVMessage,
 };
 use tokio::sync::mpsc;
