+++ anchor/qbft_manager/src/lib.rs
@@ -12,8 +12,8 @@ use ssv_types::{
     Cluster, CommitteeId, OperatorId as QbftOperatorId, OperatorId,
     consensus::{BeaconVote, QbftData, ValidatorConsensusData},
     domain_type::DomainType,
+    message::SignedSSVMessage,
     msgid::{DutyExecutor, MessageId, Role},
-    signed_message::SignedSSVMessage,
 };
 use tokio::{
     sync::{
